pyproject.toml
[tool.poetry]
name = "swat4stats"
version = "0.1.0"
description = ""
authors = ["Sergei Khoroshilov <kh.sergei@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.1.7"
celery = "5.2.7"
bleach = "6.0.0"
requests = "2.28.2"
pytz = "2023.3"
Markdown = "3.4.3"
psycopg2-binary = "2.9.5"
redis = "4.5.4"
raven = "6.10.0"
uWSGI = "2.0.21"
python-dateutil = "2.8.2"
ipwhois = "1.2.0"
django-cacheback = "3.0.0"
django-debug-toolbar = "3.8.1"
django-countries = "7.5.1"
django-filter = "23.1"
djangorestframework = "3.14.0"
voluptuous = "0.13.1"
coverage = "7.2.2"
pytest = "7.2.2"
pytest-django = "4.5.2"
pytest-localserver = "0.7.1"
pytest-cov = "4.0.0"
Faker = "18.3.1"
factory-boy = "3.2.1"
aiohttp = "3.8.4"
django-redis = "5.2.0"
pytest-repeat = "0.9.1"

[tool.poetry.group.dev.dependencies]
flake8 = "6.0.0"
pyupgrade = "3.3.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings"
norecursedirs = ".git"
python_files = "test_*.py"
addopts = "--tb=native --reuse-db"
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
